ci: upload artifacts to release on tag event
release-drafter creates draft releases. Even when publishing release-drafter
and thus a github action is the owner of the release. Hence a build is not
triggered after publishing the release. However, when publishing the draft
the tag is created by a person and we can use this to trigger the build action